% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_population}
\alias{create_population}
\title{Create a mutationtree population}
\usage{
create_population(size, base_fitness)
}
\arguments{
\item{size}{The size of the population}

\item{base_fitness}{The base fitness of genotypes in the population}
}
\value{
A mutationtree population
}
\description{
\code{create_population} creates a mutationtree population. The population is
started with one genotype at the root of the tree with \code{size}
individuals and fitness \code{base_fitness}.
}

