% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dilute}
\alias{dilute}
\title{Dilute a population}
\usage{
dilute(population, dfactor)
}
\arguments{
\item{population}{A mutationtree population}

\item{dfactor}{The dilution factor, a value in the range [0,1]}
}
\value{
A diluted population
}
\description{
\code{dilute} thins a population by a given factor.
}

