% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mutation_tree.R
\name{draw_mutation_tree}
\alias{draw_mutation_tree}
\title{Draw a mutation annotated tree}
\usage{
draw_mutation_tree(
  tree_and_sequences_asr,
  file = NULL,
  width = 6,
  height_per_sequence = 0.03,
  lines = list(),
  modified_tip_labels = list(),
  modified_node_labels = list(),
  additional_node_labels = list(),
  tip_text_cex = 0.2,
  tip_text_label_color = "grey70",
  tip_text_nudge = c(0.002, 0),
  tip_text_adj = c(0, 0.5),
  node_text_cex = 0.2,
  node_text_nudge = c(-0.001, -0.001),
  node_text_adj = c(1, 1),
  tip_node_text_aa_mutation_color = "grey30",
  tip_node_text_aa_mutation_special_colors = list(red = c(145, 155, 156, 158, 159, 189,
    193), orange = c(44L, 45L, 46L, 47L, 48L, 50L, 51L, 53L, 54L, 57L, 59L, 62L, 63L,
    67L, 75L, 78L, 80L, 81L, 82L, 83L, 86L, 87L, 88L, 91L, 92L, 94L, 96L, 102L, 103L,
    109L, 117L, 121L, 122L, 124L, 126L, 128L, 129L, 130L, 131L, 132L, 133L, 135L, 137L,
    138L, 140L, 142L, 143L, 144L, 145L, 146L, 150L, 152L, 155L, 156L, 157L, 158L, 159L,
    163L, 165L, 167L, 168L, 170L, 171L, 172L, 173L, 174L, 175L, 176L, 177L, 179L, 182L,
    186L, 187L, 188L, 189L, 190L, 192L, 193L, 
     194L, 196L, 197L, 198L, 201L, 203L,
    207L, 208L, 209L, 212L, 213L, 214L, 215L, 216L, 217L, 218L, 219L, 226L, 227L, 228L,
    229L, 230L, 238L, 240L, 242L, 244L, 246L, 247L, 248L, 260L, 261L, 262L, 265L, 273L,
    275L, 276L, 278L, 279L, 280L, 294L, 297L, 299L, 300L, 304L, 305L, 307L, 308L, 309L,
    310L, 311L, 312L)),
  tip_node_text_syn_mutation_color = "grey70",
  line_width = 0.3,
  line_color = "grey50",
  line_text_color = "black",
  line_text_cex = 1,
  line_text_adj = c(1, 1),
  line_text_nudge = c(-0.01, -0.003),
  node_addtext_cex = node_text_cex * 4,
  node_addtext_color = "black",
  node_addtext_nudge = c(-0.004, -0.006),
  node_addtext_adj = c(1, 1),
  node_point_color = "black",
  node_point_cex = 0.8,
  edge_color = "grey80",
  edge_width = NULL,
  x_lim_expand = c(0.02, 0.2),
  y_lim_expand = c(0.02, 0.02),
  node_nums = F
)
}
\arguments{
\item{tree_and_sequences_asr}{...}

\item{file}{...}

\item{width}{...}

\item{height_per_sequence}{...}

\item{lines}{...}

\item{modified_tip_labels}{...}

\item{modified_node_labels}{...}

\item{additional_node_labels}{...}

\item{tip_text_cex}{...}

\item{tip_text_label_color}{...}

\item{tip_text_nudge}{...}

\item{tip_text_adj}{...}

\item{node_text_cex}{...}

\item{node_text_nudge}{...}

\item{node_text_adj}{...}

\item{tip_node_text_aa_mutation_color}{...}

\item{tip_node_text_aa_mutation_special_colors}{...}

\item{tip_node_text_syn_mutation_color}{...}

\item{line_width}{...}

\item{line_color}{...}

\item{line_text_color}{...}

\item{line_text_cex}{...}

\item{line_text_adj}{...}

\item{line_text_nudge}{...}

\item{node_addtext_cex}{...}

\item{node_addtext_color}{...}

\item{node_addtext_nudge}{...}

\item{node_addtext_adj}{...}

\item{node_point_color}{...}

\item{node_point_cex}{...}

\item{edge_color}{...}

\item{edge_width}{...}

\item{x_lim_expand}{...}

\item{y_lim_expand}{...}

\item{node_nums}{...}
}
\value{
...
}
\description{
Draw a mutation annotated tree
}
