% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evolve}
\alias{evolve}
\title{Evolve a population for a number of generations}
\usage{
evolve(population, generations = 1, start = 1, population_size,
  mutation_rate, progress_bar = TRUE)
}
\arguments{
\item{population}{A mutationtree population}

\item{generations}{The number of generations to evolve (positive integer, default: 1)}

\item{start}{Start generation number (default: 1)}

\item{population_size}{The size of the resulting population. Since reproduction is
done stochastically, the resulting population may not be this exact value.}

\item{mutation_rate}{Rate at which mutations arise}

\item{progress_bar}{Whether or not to show a progress bar (default: True)}
}
\value{
A modified mutationtree population
}
\description{
\code{evolve} subjects the given population to a number of rounds of evolution. During this process,
the population first reproduces via \code{\link{reproduce}}, creating the next generation.
}
\note{
This function simply calls \code{\link{reproduce}},
\code{\link{mutate}}, and then \code{\link{prune}} on the population for the
specified number of cycles (generations).
}
\seealso{
\code{\link{reproduce}}

\code{\link{mutate}}

\code{\link{prune}}
}

