% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_between_zeroone}
\alias{is_between_zeroone}
\title{Return whether a given value is between 0 and 1, inclusive}
\usage{
is_between_zeroone(x)
}
\arguments{
\item{x}{Some numeric value (scalar)}
}
\value{
\code{TRUE} if \code{x} is a single numeric value in the range [0,1],
and \code{FALSE} otherwise.
}
\description{
\code{is_between_zerooone} returns a boolean value indicating whether (TRUE)
or not (FALSE) the given value is numeric and between 0 and 1, inclusive.
}

