% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mutate}
\alias{mutate}
\title{Mutate individuals in the population}
\usage{
mutate(population, mu)
}
\arguments{
\item{population}{A mutationtree population}

\item{mu}{The mutation rate}
}
\value{
A modified mutationtree population
}
\description{
\code{mutate}
}

