% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print_population}
\alias{print_population}
\title{Print information about the mutationtree population}
\usage{
print_population(population)
}
\arguments{
\item{population}{A mutationtree population}
}
\description{
\code{print_population} prints some basic information about the population
}

