% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prune}
\alias{prune}
\title{Remove extinct leaf nodes from the population graph}
\usage{
prune(population)
}
\arguments{
\item{population}{A mutationtree population}
}
\value{
A pruned mutationtree population
}
\description{
To help minimize the size of a graph, \code{prune} removes nodes that both
have zero abundance and are leaves (i.e., have no children). This second
condition is added so that extant lineages with extinct parents aren't
pruned. While more aggressive pruning is perhaps possible, this should remove
most unneeded nodes over time.
}

