% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reproduce}
\alias{reproduce}
\title{Grow the population, reproducing to given capacity}
\usage{
reproduce(population, popsize)
}
\arguments{
\item{population}{A mutationtree population}

\item{popsize}{The size of the resulting population. Since reproduction is
done stochastically, the resulting population may not be this exact value.}

\item{A}{modified mutationtree population}
}
\description{
\code{reproduce} allows the population to grow to the given capacity.
Reproduction is done using fitness-proportional selection based on the
fitness associated with each genotype.
}

